Refactor onboarding, reuse most of the code for login and register activities
