+++ app/src/main/res/layout/activity_registration.xml
@@ -8,11 +8,11 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.olegmisko.newsapplication.main.Activities.RegistrationActivity">
+    tools:context="com.olegmisko.newsapplication.main.Activities.OnboardingActivity">
 
 
     <ImageView
-        android:id="@+id/imageView"
+        android:id="@+id/onboarding_icon"
         android:layout_width="150dp"
         android:layout_height="150dp"
         android:layout_centerHorizontal="true"
@@ -35,7 +35,7 @@
         android:id="@+id/input_layout_username"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@id/imageView"
+        android:layout_below="@id/onboarding_icon"
         android:layout_centerHorizontal="true"
         android:paddingTop="10dp">
