+++ build.gradle
@@ -5,7 +5,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
+        classpath 'com.android.tools.build:gradle:2.3.0'
         classpath "io.realm:realm-gradle-plugin:2.3.0"
 
         // NOTE: Do not place your application dependencies here; they belong
